services/ai-service/requirements.txt
fastapi
uvicorn
requests
pydantic 
dotenv
celery==5.4.0
redis==5.0.8
google-generativeai
